backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic-settings==2.6.0
pymupdf==1.24.13
chromadb==0.5.23
sentence-transformers==3.3.1
openai==1.57.4
numpy==1.26.4
python-dotenv==1.0.1
